Created individual model files
