+++ client/src/app/@core/data/maps.service.ts
@@ -2,11 +2,6 @@ import {Injectable} from '@angular/core';
 import {Observable} from 'rxjs';
 import {HttpClient} from '@angular/common/http';
 
-export interface MomentumMap {
-  id: number;
-  name: string;
-}
-
 @Injectable()
 export class MapsService {
