+++ client/src/app/@core/models/activity.model.ts
@@ -1,0 +1,5 @@
+export interface Activity {
+  id: number;
+  type: number;
+  data: string;
+}
