+++ client/src/app/@core/models/momentum-map.model.ts
@@ -1,0 +1,4 @@
+export interface MomentumMap {
+  id: number;
+  name: string;
+}
