+++ client/src/app/@core/models/profile.model.ts
@@ -1,0 +1,9 @@
+export interface UserProfile {
+  id: string;
+  alias: string;
+  avatarURL: string;
+  bio?: string;
+  twitterName?: string;
+  discordName?: string;
+  youtubeName?: string;
+}
