+++ client/src/app/@core/models/user.model.ts
@@ -1,0 +1,9 @@
+import {UserProfile} from './profile.model';
+
+export interface User {
+  id: string;
+  permissions: number;
+  createdAt?: string;
+  updatedAt?: string;
+  profile: UserProfile;
+}
